moved helper function from EditTextInputAction.kt to TextInputAction.kt and also changed the access of function .
