+++ testing/build.gradle
@@ -52,9 +52,9 @@ dependencies {
       'androidx.lifecycle:lifecycle-livedata-ktx:2.2.0-alpha03',
       'androidx.test.espresso:espresso-core:3.2.0',
       'androidx.test:runner:1.2.0',
+      'com.google.android.material:material:1.2.0-alpha02',
       'com.google.dagger:dagger:2.24',
       'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.2.2',
-      'com.google.android.material:material:1.2.0-alpha02',
       'org.robolectric:robolectric:4.3',
       'org.jetbrains.kotlin:kotlin-reflect:$kotlin_version',
       project(":domain"),
